### 2024-05-30 #6812 @editorialbot
[PRE REVIEW]: SONIC: Software for Optical Navigation and Instrument Calibration
**Submitting author:** @athrasher7 (<a href="http://orcid.org/0009-0006-6019-1816">Ava Thrasher</a>)
**Repository:** https://github.com/opnavlab/sonic
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
