### 2024-06-21 #6812 @Subho07
Hi Prashant,

Thanks for your email. However, I will not be able to review this time.

Best regards,
Subhadip

-------------------------------------
Subhadip Dey
Assistant Professor,
Department of Agricultural and Food Engineering,
Indian Institute of Technology Kharagpur,
Kharagpur, India - 721302
*Mail To:* ***@***.*** ***@***.***>*
*Contact:* 03222-284540

On Fri, Jun 21, 2024 at 10:17 PM Prashant K. Jha ***@***.***>
wrote:

> Howdy! @Subho07 <https://github.com/Subho07>, @Muhao-Chen
> <https://github.com/Muhao-Chen>, @athulpg007
> <https://github.com/athulpg007>, and @scivision
> <https://github.com/scivision>, would you be interested in reviewing the
> JOSS submission above (pdf link
> <https://github.com/openjournals/joss-papers/blob/joss.06812/joss.06812/10.21105.joss.06812.pdf>)?
> The corresponding software is at this link
> <https://github.com/opnavlab/sonic>.
>
> This link <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>
> covers almost everything about the review process; I will be happy to help
> with more information.
>
> Please let me know if you are interested. Thank you!
>
