### 2024-06-21 #6812 @prashjha
Hello @athrasher7, we now have two reviewers. The review will take place in a new issue. Meanwhile, please fix the 'Missing DOIs' problem with your JOSS draft and keep an eye on reviewer's comments or any issue they raise in your software repository.
