### 2024-06-21 #6916 @editorialbot
[REVIEW]: SONIC: Software for Optical Navigation and Instrument Calibration
**Submitting author:** @athrasher7 (<a href="http://orcid.org/0009-0006-6019-1816">Ava Thrasher</a>)
**Repository:** https://github.com/opnavlab/sonic
**Branch with paper.md** (empty if default branch): Docs
**Version:** SONICV0
