### 2024-06-23 opnavlab/sonic#6 @Muhao-Chen
JOSS review general comments
> @athrasher7 Below are my review comments for SONIC [openjournals/joss-reviews#6812](https://github.com/openjournals/joss-reviews/issues/6812)
>
> * [x] Please add a `setup.m` function and use the addpath command to automatically add all the main and subfolders to the main folder.
>
> * [x] Please add a readme.md to all the folders so that people can read easily what's going on in that folder.
>
> * [x] Please add a one-sentence description of the licenses to each code file.
>
> * [x] Please describe how people can contribute to this software package on the first page. Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
>
> * [x] In my personal project [TsgFEM](https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM) I use a MATLAB script called `test_all.m` (https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/blob/main/Software_Verification_and_Examples/test_all.m) to run all tests/demos automatically. Consider adopting a similar automated testing scheme.
>
> * [x] Please add more introduction to how to use the software to solve real-world analysis problems on the first page and the paper. It is preferred to have a few bullet points.
>
> * [x] Other languages are used apart from Matlab and HTML. Consider writing a paragraph to clarify if related software or environment setup is needed.
>
> * [x] Versions of MATLAB toolbox dependencies (Image Processing Toolbox, Computer Vision System Toolbox) and computer requirements need to be explained in more detail.
>
