### 2024-06-23 #6916 @Muhao-Chen
@athrasher7, thank you for submitting this interesting work. I have downloaded and tested the software, and I have a few comments that could benefit from some changes. Please refer to the details I've highlighted in this issue: https://github.com/opnavlab/sonic/issues/6

The major points concern automated testing, licenses, community guidelines, and the README. I have also provided some suggestions on these issues.
