### 2024-06-28 #6812 @athrasher7
@prashjha Thanks for being patient, I've been on leave. I will fix the missing dois for the papers, but what should I do for the references which are repositories (or in the case of GABLE, a MATLAB file exchange)? In the example paper some of the citations don't have dois so I thought a url would be sufficient. Thanks!
