### 2024-06-28 6c882e3
Update paper.bib
Added more DOIs that I could find. Those left with no DOI didn't have an associated one that I could find.
