+++ paper/paper.bib
@@ -52,7 +52,7 @@
   year = {2010},
   volume = {139},
   pages = {1782-1800},
-  doi = {10.1088/0004-6256/139/5/1782},
+  doi = {10.1088/0004-6256/139/5/1782}
 }
 
 @article{sangwine2016, 
@@ -92,6 +92,7 @@
   title     = {Multiple View Geometry in Computer Vision},
   year      = {2003},
   publisher = {Cambridge University Press},
+  doi       = {10.1017/CBO9780511811685}
 }
 
 @book{semple1952,
@@ -99,6 +100,7 @@
   title     = {Algebraic Projective Geometry},
   year      = {1952},
   publisher = {Oxford University Press},
+  doi       = {10.2307/3608975}
 }
 
 @book{richtergebert2011,
@@ -106,11 +108,12 @@
   title     = {Perspective on Projective Geometry: A Guided Tour Through Real and Complex Geometry},
   year      = {2011},
   publisher = {Springer},
+  doi       = {10.1007/978-3-642-17286-1}
 }
 
 @book{lengyel2024,
   author    = {E. Lengyel},
-  title     = {Projective Geometry Illuminated},
+  title     = {Projective Geometry Algebra Illuminated},
   year      = {2024},
   publisher = {Terathon Software, LLC},
 }
