### 2024-06-28 5115a85
Create readme.md
Adding readme to examples folder
