+++ +examples/readme.md
@@ -1,0 +1,7 @@
+The +examples folder contains MATLAB live scripts which each have demonstrations of SONIC use cases. 
+These examples are not comprehensive, but show the general work flow of using the SONIC building 
+blocks to solve more complex problems.
+
+- **ScanLinesTutorial.mlx** Demonstrates the generic ability to extract image values along a specified scanline, which is useful for horizon-based OPNAV.
+- **SyntheticStarImgTutorial.mlx** Demonstrates creating a synthetic image of the star Arcturus, and its surrounding starfield in SONIC. Being able to construct an expected image is an essential operation for OPNAV.
+- **TriangVestaReconTutorial.mlx** Illustrates the use of the triangulation in SONIC to perform 3D structure reconstruction on the asteroid Vesta from features extracted in images. A few different approaches to triangulation 
