### 2024-06-28 a2178bd
Create readme.md
Adding readme to +examples/+data
