+++ +examples/+data/readme.md
@@ -1,0 +1,1 @@
+The +exmaples/+data directory contains data files that the MATLAB live scripts reference. The tutorials will need these files to run correctly.
