### 2024-06-28 7461a8b
Create readme.md
Adding readme to +sonic folder
