+++ +sonic/readme.md
@@ -1,0 +1,1 @@
+The +sonic folder contains all of SONIC's classes. For more information about individual classes and their properties and methods, please visit the [API documentation](https://opnavlab.github.io/sonic/).
