### 2024-06-28 3b4da9a
Create readme.md
Adding readme to +sonic/+data.
