+++ +sonic/+data/readme.md
@@ -1,0 +1,1 @@
+The +sonic/+data folder contains versions of the Hipparcos catalog and constellation information which has been specifically organized for use with SONIC. 
