+++ paper/readme.md
@@ -1,0 +1,1 @@
+Paper content and citations for the Journal of Open Sourced Software article submission.
