+++ docs_utils/readme.md
@@ -1,0 +1,1 @@
+The docs_utils folder contains scripts for generating the neccesary html files for the online API documentation. None of these scripts are necessary for using the SONIC package.
