### 2024-06-28 8bd62e7
Update README.md
Adding more direct list of real-world navigation problems that SONIC could be used to help solve.
