### 2024-06-28 opnavlab/sonic#6 @athrasher7
In regards to adding more about solving real-world problems, I've added a bullet point list in the readme, but the paper already includes a list in the first paragraph:

"Finally, SONIC offers solutions to the commonly encountered problems of star identification [@mortari2004][@lang2010], triangulation [@henry2023], Horizon-based OPNAV [@christian2021], and more."

These three are very real-world navigation problems. Is there something more you are looking for here? Thanks!
