### 2024-07-01 opnavlab/sonic#7 @athulpg007
JOSS review comments
Paper Comments:

- The paper is very-well written, and the code is well-documented.
- My recommendation is to include at least a couple of figures in the paper which illustrate the capability of the SONIC package to real world problems. Examples include: 1) the Synthetic Star Image of Arcturus, and its surrounding starfield, 2) smear removal of the Bennu asteroid image, 3) intensity of the image over a scan line, 4) Vesta 3D reconstruction. Including some of these figures in the paper, in my opinion would enrich the quality, and offer a preview of the capabilities to a diverse, non-specialist audience reading the paper.

Code Comments:
- The code is generally well-written, and well-organized. The examples in the documentation clearly illustrate the applications.
- There are no unit tests for any of the methods in +sonic. Please consider adding some unit tests for at least some of the core methods, or provide a justification how the authors verified the functions used are implemented correctly.
- There are no integration tests which exercise multiple methods used in a workflow. Please consider adding some integration tests, or alternatively provide a justification (for example: can the software reproduce some published results elsewhere?)
- "Note: As of 5/16/2024, SONIC has a few MATLAB toolbox dependencies which will require installation prior to full operation."
  - [optional] Instead of a hard-coded date, consider creating a tagged release, and use the version tag here instead (eg: v1.0)
  - [optional] consider replacing "SONIC has a few MATLAB toolbox dependencies" with "SONIC requires the following MATLAB toolboxes installed."
  - please specify the minimum required versions for the two dependencies, and if the latest versions are preferred.
- There are no community guidelines for contributions. Please consider adding a minimal CONTRIBUTING.md: [https://github.com/astropy/astropy?tab=readme-ov-file#contributing-code-documentation-or-feedback](https://github.com/astropy/astropy?tab=readme-ov-file#contributing-code-documentation-or-feedback)

Minor comments:
- The first reference appears it should be "Astrometry.net", instead of "Astronomy.net"
