### 2024-07-01 #6916 @athulpg007
I have completed my initial review of the submission. Please see the comments in this issue: https://github.com/opnavlab/sonic/issues/7.
