### 2024-07-02 61a5e5c
adding license descriptions to each file
