+++ +sonic/DistortionModel.m
@@ -1,3 +1,5 @@
+% This software is made available under the MIT license. See SONIC/LICENSE
+% for details.
 classdef (Abstract) DistortionModel
     methods (Abstract)
         pointsd = distort(obj,points)
