### 2024-07-02 d4d6a87
Removing documentation from main branch
