+++ docs_utils/index.rst
@@ -1,19 +1,0 @@
-
-Software for Optical Navigation and Instrument Calibration (SONIC) Documentation
-=================================
-Please see below for instructions on installation, examples of using SONIC, and a listing of documentation for the classes of SONIC (api).
-
-.. toctree::
-   :maxdepth: 1
-
-   install
-   examples
-   api
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
