+++ docs_utils/api_src/planes3.rst
@@ -1,10 +1,0 @@
-Planes3
-==========================
-
-.. automodule::  +sonic
-    
-
-.. autoclass:: Planes3
-	:show-inheritance:
-	:members:
-	:undoc-members:
