+++ docs/readme.md
@@ -1,1 +1,0 @@
-Files generated for the API documentation. These files are generated by the scripts in docs_utils and are then used to populate the online documentation for SONIC.
