+++ docs_utils/api_src/distortionmodel.rst
@@ -1,10 +1,0 @@
-DistortionModel
-==========================
-
-.. automodule::  +sonic
-    
-
-.. autoclass:: DistortionModel
-	:show-inheritance:
-	:members:
-	:undoc-members:
