+++ docs_utils/api_src/hipparcos.rst
@@ -1,10 +1,0 @@
-Hipparcos
-==========================
-
-.. automodule::  +sonic
-    
-
-.. autoclass:: Hipparcos
-	:show-inheritance:
-	:members:
-	:undoc-members:
