+++ docs_utils/api_src/centroider.rst
@@ -1,10 +1,0 @@
-Centroider
-==========================
-
-.. automodule::  +sonic
-    
-
-.. autoclass:: Centroider
-	:show-inheritance:
-	:members:
-	:undoc-members:
