+++ docs_utils/api_src/aberration.rst
@@ -1,10 +1,0 @@
-Aberration
-==========================
-
-.. automodule::  +sonic
-    
-
-.. autoclass:: Aberration
-	:show-inheritance:
-	:members:
-	:undoc-members:
