### 2024-07-02 ae5f269
Update README.md
adding versions to the toolbox dependencies
