+++ README.md
@@ -30,8 +30,8 @@ for demonstrative purposes:
 
 Note: As of 5/16/2024, SONIC has a few MATLAB toolbox dependencies which will
 require installation prior to full operation including:
-- Image Processing Toolbox
-- Computer Vision System Toolbox
+- Image Processing Toolbox v 24.1
+- Computer Vision System Toolbox version 24.1
 
 # Feedback
 If you have a suggestion, please use the GitHub [Issues](https://github.com/opnavlab/sonic/issues) tab to let us know!
