### 2024-07-02 opnavlab/sonic#6 @athrasher7
Following up on the rest of the items:
1. SONIC was specifically designed to be used without a set up function. All of the internal folders are '+' folders, which means they are automatically added to the sonic path. The user only needs to add "sonic/" to the path of their working directory, which is a one-liner in MATLAB.
4. We are in contact with Georgia Tech's OTL about third party contributions. Once we hear back we will include a statement on this.
5. Thank you for the suggestion, but the SONIC demos are all Matlab live scripts, so a test script to run them wouldn't be intuitive. The user should run these demos locally and then compare the output to the pre-run demos in the API documentation to verify their installation. I've added a sentence to the readme stating this, and to the documentation.
7. The only languages should now be matlab and markdown. Previously there was more because of python scripts to customize the documentation. For organization, all documentation utilities were moved to a separate branch called Docs.
8. Version info was added.

Thanks again for the review! I will let you know when we add more info on third party contributions.
