### 2024-07-05 a437bdf
Update paper.md
Adding images as example usage of SONIC, per reviewers request.
