### 2024-07-05 c02f5fd
Update paper.bib
fixing astrometry citation
