+++ paper/paper.bib
@@ -38,7 +38,7 @@
 }
 
 @misc{astronet,
-  title = {Astronomy.net},
+  title = {astrometry.net},
   year = {2015},
   publisher = {GitHub},
   journal = {GitHub repository},
