### 2024-07-05 opnavlab/sonic#7 @athrasher7
Thank you for your review!

Paper Comments:

- I have added three figures to the paper that I think are useful for a new reader who hasn't looked at our repository yet. Each picture has a caption highlighting what was done (despite reading documentation, I'm not confident on how the captions are formatted). I included a new paragraph encouraging readers to run the examples locally to see how they work.

Code comments:

- We did much of our unit testing and general testing in a private repo to avoid cluttering the public repo. We would rather not upload these for the main reason that there are so many, and it would just be distracting for the average user. Where would you recommend that we add information on how unit testing was conducted, and what information are you looking for specifically?
- The workflows provided have references included for the user to check our implementation for themselves - the math and approach is intentionally transparent. There are many integrated capabilities, and a lot of them will be unique to a use case. Because of this, I'm not sure of an efficient way to show integration testing beyond providing the examples we did. Are there specific interactions you would recommend we include, or more information than the references we should add?
- The time tag was removed, and the versions used for development were added. We can't guarantee older or newer versions will work at this time. I changed the wording to your suggestion as well in the readme.
- We are in contact with our office about their stance on contributions. We will upload information to the readme once we hear back.

Minor Comments:

- Fixed
