### 2024-07-05 3e3991d
normalizing to unit vector before calling PointsS2
