### 2024-07-05 861d700
fixing normalization
