### 2024-07-06 opnavlab/sonic#7 @athulpg007
Thank you for the updates.

- Please re-make the PDF using @editorialbot generate pdf in https://github.com/openjournals/joss-reviews/issues/6916. I tried to do it, but the workflow ran into a error.
- It is considered good practice to have unit tests for the methods in a tests/ directory at the root. This allows for a regression test suite to ensure future changes/contributions do not break existing functionality. If you prefer to not have the unit tests, that is still acceptable. Please provide an explanation as to how internal users/contributors can verify they do not introduce bugs in the future. One suggestion would be to "run all the examples locally and compare the outputs to the pre-run demos in the SONIC API documentation" if any changes are made, plus "add an example illustrating any new functionality".
- The workflows provided in the examples are sufficient for integration tests.

I will re-review the paper once the PDF is available, and you hear back about contribution guidelines.
