+++ docs_utils/api_src/camera.rst
@@ -1,0 +1,10 @@
+Camera
+==========================
+
+.. automodule::  +sonic
+    
+
+.. autoclass:: Camera
+	:show-inheritance:
+	:members:
+	:undoc-members:
