+++ docs_utils/api_src/scanlines.rst
@@ -1,0 +1,10 @@
+ScanLines
+==========================
+
+.. automodule::  +sonic
+    
+
+.. autoclass:: ScanLines
+	:show-inheritance:
+	:members:
+	:undoc-members:
