+++ docs_utils/api_src/pointss2.rst
@@ -1,0 +1,10 @@
+PointsS2
+==========================
+
+.. automodule::  +sonic
+    
+
+.. autoclass:: PointsS2
+	:show-inheritance:
+	:members:
+	:undoc-members:
