+++ docs_utils/examples.rst
@@ -1,0 +1,10 @@
+Examples with SONIC
+=================================
+
+Say some things here about examples
+
+Link to an Open With MATLAB Online version of the script below!
+
+Link to the live script example, evaluated and exported as HTML:
+
+`This is the link <_static/SyntheticStarImgTutorial.html>`_
