+++ docs_utils/api_src/tolerances.rst
@@ -1,0 +1,10 @@
+Tolerances
+==========================
+
+.. automodule::  +sonic
+    
+
+.. autoclass:: Tolerances
+	:show-inheritance:
+	:members:
+	:undoc-members:
