+++ docs_utils/api.rst
@@ -1,0 +1,38 @@
+API
+===
+.. toctree::
+	./api_src/aberration
+	./api_src/attitude
+	./api_src/brownconrady
+	./api_src/camera
+	./api_src/centroider
+	./api_src/conic
+	./api_src/constants
+	./api_src/distortionmodel
+	./api_src/edgefinder
+	./api_src/ellipsoid
+	./api_src/geometryp2
+	./api_src/geometryp3
+	./api_src/hipparcos
+	./api_src/horizonopnav
+	./api_src/image
+	./api_src/kvector
+	./api_src/lines2
+	./api_src/lines3
+	./api_src/math
+	./api_src/meetjoinable
+	./api_src/pinhole
+	./api_src/planes3
+	./api_src/points2
+	./api_src/points3
+	./api_src/pointss2
+	./api_src/project
+	./api_src/quadric
+	./api_src/samplegeom2d
+	./api_src/scanlines
+	./api_src/spherecoords
+	./api_src/starcatalog
+	./api_src/starid
+	./api_src/tolerances
+	./api_src/triangulation
+	./api_src/units
