+++ docs_utils/api_src/aberration.rst
@@ -1,0 +1,10 @@
+Aberration
+==========================
+
+.. automodule::  +sonic
+    
+
+.. autoclass:: Aberration
+	:show-inheritance:
+	:members:
+	:undoc-members:
