+++ docs_utils/api_src/points2.rst
@@ -1,0 +1,10 @@
+Points2
+==========================
+
+.. automodule::  +sonic
+    
+
+.. autoclass:: Points2
+	:show-inheritance:
+	:members:
+	:undoc-members:
