+++ docs_utils/api_src/horizonopnav.rst
@@ -1,0 +1,10 @@
+HorizonOPNAV
+==========================
+
+.. automodule::  +sonic
+    
+
+.. autoclass:: HorizonOPNAV
+	:show-inheritance:
+	:members:
+	:undoc-members:
