+++ docs_utils/api_src/geometryp2.rst
@@ -1,0 +1,10 @@
+GeometryP2
+==========================
+
+.. automodule::  +sonic
+    
+
+.. autoclass:: GeometryP2
+	:show-inheritance:
+	:members:
+	:undoc-members:
