+++ docs_utils/api_src/ellipsoid.rst
@@ -1,0 +1,10 @@
+Ellipsoid
+==========================
+
+.. automodule::  +sonic
+    
+
+.. autoclass:: Ellipsoid
+	:show-inheritance:
+	:members:
+	:undoc-members:
