+++ docs_utils/index.rst
@@ -1,8 +1,21 @@
-
-Welcome to SONIC's documentation!
+SONIC API Documentation
 =================================
 
-We can put some more welcome text here!
+This open-sourced, object oriented MATLAB package is a toolkit built on years of research in
+optical navigation (OPNAV), with a foundation in principled projective geometry. Many 
+capabilities needed for space-flight OPNAV are made transparent and configurable within the SONIC classes, 
+with minimal assumption of the analyst's intentions. This allows each user to construct their own 
+unique workflows for a specific research or analysis problems. SONIC was designed to help simplify navigation solutions such as (but not limited to):
+
+- Horizon-Based OPNAV
+- Triangulation
+- Star Identification
+- 3D-Reconstruction
+- And more
+
+SONIC was created by the researchers and students of the [Space Exploration and Design 
+Laboratory](https://seal.ae.gatech.edu/) at the Georgia Institute of Technology for student, 
+academic researchers, space science professionals, or anyone curious to learn more about OPNAV.
 
 .. toctree::
    :maxdepth: 1
