### 2024-07-12 cbfafc2
Create CONTRIBUTING.md
Adding some basic information about contributing
