+++ CONTRIBUTING.md
@@ -1,0 +1,29 @@
+# Contribution Guidelines
+
+Thank you for considering contributing to Software for Optical Navigation and Instrument Calibration (SONIC). 
+SONIC is an open sourced software, and the developers are happy to receive feedback and suggestions regarding documentation, feature requests, and bug reports.
+Please read the guidelines below to make sure your concerns or suggestions are addressed by the developers. 
+
+# How to report a bug
+If you find a bug while using SONIC, please open an issue under the Issues tab.
+
+When filing an issue, make sure to answer these questions:
+1. What version of MATLAB are you using?
+2. What did you do?
+3. What did you expect to see?
+4. What did you see instead?
+
+If you know the cause of the bug, feel free to include that as well! A SONIC developer will acknowledge your issue and update the issue as it is investigated.
+
+# How to suggest a feature or change
+If you want to see a new feature in SONIC, or think that a current feature could be improved, please submit an issue under the Issues tab. Specifically, let us know:
+1. What new feature do you want to see, or what feature currently do you want to see improved?
+2. How would this addition/change aid in the development of optical navigation (OPNAV) solutions?
+3. How would you suggest we implement the new feature/change?
+
+
+# Code review process
+The research team at SEAL is comprised of graduate students and full time research engineers, so availability of resources to respond to issues may vary throughout the year and based on the scope of the issue. 
+At a minimum, issues will be checked and responded to weekly to give an estimated time-needed to complete a fix for the issue.
+For each new feature or large change, a lab code review meeting is held to approve the changes before integrating them into SONIC. 
+
