+++ README.md
@@ -34,8 +34,8 @@ Note: SONIC requires the following MATLAB toolboxes installed:
 - Image Processing Toolbox version 24.1
 - Computer Vision System Toolbox version 24.1
 
-# Feedback
-If you have a suggestion, please use the GitHub [Issues](https://github.com/opnavlab/sonic/issues) tab to let us know!
+# Contributing
+If you are interesting in contributing, have a feature request, or have found a bug in SONIC, please checkout [CONTRIBUTING.md](https://github.com/opnavlab/sonic/blob/main/CONTRIBUTING.md) for more information.
 
 # References
 Please see [REFERENCES.md](https://github.com/opnavlab/sonic/blob/main/REFERENCES.md) for a list of references used in the development of SONIC.
