+++ docs_utils/examples.rst
@@ -1,10 +1,8 @@
 Examples with SONIC
 =================================
 
-Say some things here about examples
+As demonstration of a few applications of SONIC to real world OPNAV problem, the following example MATLAB live scripts were created.
 
-Link to an Open With MATLAB Online version of the script below!
-
-Link to the live script example, evaluated and exported as HTML:
-
-`This is the link <_static/SyntheticStarImgTutorial.html>`_
+- `Create Scan Lines in Image <_static/ScanLinesTutorial.html>`_: Demonstrates the generic ability to extract image values along a specified scanline, which is useful for horizon-based OPNAV.
+- `Generate Synthetic Star Image <_static/SyntheticStarImgTutorial.html>`_: Demonstrates creating a synthetic image of the star Arcturus, and its surrounding starfield in SONIC. Being able to construct an expected image is an essential operation for OPNAV.
+- `3D Reconstruction of Vesta <_static/TriangVestaReconTutorial.html>`_: Illustrates the use of the triangulation in SONIC to perform 3D structure reconstruction on the asteroid Vesta from features extracted in images. A few different approaches to triangulation are provided in SONIC, which can be used both for estimating the spacecraft state or reconstruction.
