+++ docs_utils/index.rst
@@ -13,8 +13,8 @@ unique workflows for a specific research or analysis problems. SONIC was designe
 - 3D-Reconstruction
 - And more
 
-SONIC was created by the researchers and students of the [Space Exploration and Design 
-Laboratory](https://seal.ae.gatech.edu/) at the Georgia Institute of Technology for student, 
+SONIC was created by the researchers and students of the `Space Exploration and Design 
+Laboratory <https://seal.ae.gatech.edu/>`_ at the Georgia Institute of Technology for student, 
 academic researchers, space science professionals, or anyone curious to learn more about OPNAV.
 
 .. toctree::
