### 2024-07-12 c3ed371
Update paper.md
Trying to fix paper citations
