### 2024-07-12 5ab007b
Update paper.md
Replacing brackets. Removing them did not fix citation
