+++ paper/paper.md
@@ -33,7 +33,7 @@ Optical navigation (OPNAV) represents an increasingly common solution for spacef
 
 `SONIC` is an object-oriented MATLAB toolkit for OPNAV, designed to simplify and accelerate the production of new research results in this field. The `SONIC` toolkit provides an organized and transparent set of utilities to make this research readily available to projects looking to include OPNAV in their mission design and analysis.
 
-Projective geometry @semple1952 @richtergebert2011 and geometric algebra [@lengyel2024] is the foundation for `SONIC`. Geometric objects such as points, lines, planes, conics, and quadrics are conveniently represented in ways that enable quick analysis of their meets, joins, and projections. Image processing routines such as image scans, edge detection, estimating background noise, and centroiding are included and easily configurable depending on the desired use. `SONIC` also intuitively handles the representation of camera models and can account for optical distortions and stellar aberration. In addition, `SONIC` provides a direct interface to the Hipparcos star catalog, allowing the user to easily access and filter its entries. Finally, `SONIC` offers solutions to the commonly encountered problems of star identification [@mortari2004][@lang2010], triangulation [@henry2023], Horizon-based OPNAV [@christian2021], and more. These individual functionalities in `SONIC` can be used to synthesize larger capabilities. `SONIC` was deliberately created as a toolbox and does not include complete workflows for end-to-end capabilities. This design choice allows `SONIC` to focus on the foundational building blocks of OPNAV while making minimal assumptions about the user's application. Therefore, the user is given the freedom and flexibility to assemble the `SONIC` building blocks in their own way to produce a unique OPNAV capability or address a particular research problem. 
+Projective geometry [@semple1952] [@richtergebert2011] and geometric algebra [@lengyel2024] is the foundation for `SONIC`. Geometric objects such as points, lines, planes, conics, and quadrics are conveniently represented in ways that enable quick analysis of their meets, joins, and projections. Image processing routines such as image scans, edge detection, estimating background noise, and centroiding are included and easily configurable depending on the desired use. `SONIC` also intuitively handles the representation of camera models and can account for optical distortions and stellar aberration. In addition, `SONIC` provides a direct interface to the Hipparcos star catalog, allowing the user to easily access and filter its entries. Finally, `SONIC` offers solutions to the commonly encountered problems of star identification [@mortari2004][@lang2010], triangulation [@henry2023], Horizon-based OPNAV [@christian2021], and more. These individual functionalities in `SONIC` can be used to synthesize larger capabilities. `SONIC` was deliberately created as a toolbox and does not include complete workflows for end-to-end capabilities. This design choice allows `SONIC` to focus on the foundational building blocks of OPNAV while making minimal assumptions about the user's application. Therefore, the user is given the freedom and flexibility to assemble the `SONIC` building blocks in their own way to produce a unique OPNAV capability or address a particular research problem. 
 
 As demonstrations for applying SONIC to OPNAV problem, multiple example workflows have been provided in the SONIC repository. These cover only a few examples, which are shown in Figs. 1 through 3. For more information on how these results were obtained using SONIC, the reader is encouraged to run these examples locally.
 
