### 2024-07-12 opnavlab/sonic#7 @athrasher7
Happy Friday,

I have added a contributions file, which is pretty minimal right now, but subject to include more information about pull requests later.

I am having difficulty figuring out why the PDF will not generate. It seems like it doesn't like the inline citations, but it looks to me like they are in proper markdown format (i.e. [@refNameHere]). I've read through the documentation and can't find what's wrong. Do you have any guidance on this?

The installation page asks users to run the examples to verify their installation using the prerun examples, and the contributions file asks contributors to create their own similar example and to re run the old examples to ensure they still work.
