### 2024-07-13 opnavlab/sonic#7 @athulpg007
Could you try if this example works? If yes, please replace the image URL and caption accordingly.
```
![](https://i.imgur.com/3XPh6JY.png)
**Figure 1.** An example of a Venus aerocapture vehicle trajectory and the stagnation-point heating profile computed by using AMAT.
```
