### 2024-07-15 7f88cb8
Update paper.md
debugging figure
