### 2024-07-15 c8b47ff
Update paper.md
Debugging figure
