+++ paper/paper.md
@@ -37,8 +37,8 @@ Projective geometry [@semple1952] [@richtergebert2011] and geometric algebra [@l
 
 As demonstrations for applying SONIC to OPNAV problem, multiple example workflows have been provided in the SONIC repository. These cover only a few examples, which are shown in Figs. 1 through 3. For more information on how these results were obtained using SONIC, the reader is encouraged to run these examples locally.
 
-![](https://i.imgur.com/3XPh6JY.png)
-**Figure 1.** An example of a Venus aerocapture vehicle trajectory and the stagnation-point heating profile computed by using AMAT. 
+![](https://github.com/opnavlab/sonic/blob/main/paper/scan_lines.png)
+**Figure 1.** SONIC can be used to quickly perform a scan and do analysis on an image.. 
 
 There are a number of existing software packages that address similar problems as `SONIC`. Of particular note is the Goddard Image Analysis and Navigation Toolkit (`GIANT`), which is a well-known OPNAV API [@giant]. `GIANT` is scene-oriented, allowing the user to manipulate objects' positions and orientations within a scene, and then perform a multitude of OPNAV routines. `SONIC` seeks to complement this by design by approaching the OPNAV problem from a camera-oriented and projective geometry point of view, while also adding a few different solutions for state estimation. Other software packages have solutions for different aspects of OPNAV. For example, `GABLE` [@gable], the `Clifford Multivector Toolbox` [@sangwine2016], and NASA’s `Rigid Geometric Algebra` [@rga] all provide generic computing frameworks for geometric algebra. While `SONIC` does use concepts from geometric algebra, generic geometric algebera manipulations are not required. Furthermore, there are well-known image processing toolkits for general applications such as image processing (i.e. `OpenCV` [@opencv]) or astronomical image registration (i.e. `Tetra3` [@tetra3] and `Astrometry.net` [@lang2010][@astronet]). `SONIC` captures the intersection of principles in projective geometry, image processing, and optics needed for the growing multi-disciplinary field of OPNAV. 
 
