### 2024-07-15 66f4255
Update paper.md
Debugging figure
