### 2024-07-15 0cbe2d6
Update paper.md
Adding in fixed links for the figures
