### 2024-07-15 #6916 @athrasher7
@athulpg007 @Muhao-Chen I've fixed the paper - see the correct generation of it in the comment above. Thank you both for your review of SONIC!
