### 2024-07-25 6dd6965
reducing the quaternion norm tolerance
