### 2024-08-02 #6916 @athulpg007
@athrasher7 as the editor, @prashjha will need to formally accept the paper as the next step. Let us wait for one more week. If we do not get a response by then, we can reach out to the Associate Editor-in-Charge.
