### 2024-08-05 #6916 @prashjha
Hi @ athrasher7, for the next steps, I suggest creating a tag of your branch with the version number. Once you have created a tag, please use it to archive your code using [Zenodo](https://zenodo.org/). When you are creating a zenodo archive, ensure that the title of the archive matches the title of your JOSS paper.

Finally, let me know the version number of tag, and the DOI of zenodo archive. I will then update your JOSS paper metadata. If you have any questions, just let me know.
