### 2024-08-05 #6916 @athrasher7
@prashjha I believe I have done the Zenodo archive correctly. The tag is SONICV0, and the DOI is: 10.5281/zenodo.13222700

Please let me know if there is anything I need to fix. Thanks!
