### 2024-08-09 661e858
fixing rho min and max computation
