### 2024-08-16 #6916 @athrasher7
@prashjha Hi again, just want to ping this since I haven't heard anything this week. We are interested in getting this publication out soon, so please let me know if there is anything else we need to do to move that process forward. Thanks!
