### 2024-08-21 #6916 @prashjha
Hi @athrasher7, I read the draft, and it looks good. Can you provide the syntax to run the three examples mentioned in the draft (Figs 1, 2, and 3)?
