### 2024-08-21 #6916 @athrasher7
@prashjha I have fixed the typo in the archive title. Thanks for pointing that out! In terms of the syntax, these examples are provided in the GitHub repository as MATLAB live examples. This means that the user only needs to clone the repository, open the example files in the +example folders, and hit run to generate the results. Is this the information you are looking for?
