### 2024-08-21 #6916 @prashjha
@athrasher7, I understand that. I suggested that you update the draft with the syntax to obtain the figures in your draft.
