### 2024-08-21 2438c62
Update paper.md
Adding syntax for image generation.
