### 2024-08-21 #6916 @athrasher7
Okay! I'm not quite sure what you mean. I've added a sentence above directing the readers to clone the repository and navigate to the examples folder to find these examples, and I added which example generated the figures to each caption. Is this what you are suggesting?
