### 2024-08-25 #6916 @prashjha
> Okay! I'm not quite sure what you mean. I've added a sentence above directing the readers to clone the repository and navigate to the examples folder to find these examples, and I added which example generated the figures to each caption. Is this what you are suggesting?

@athrasher7, I suggested including the MATLAB syntax/pseudo-code required to run the examples in the figures. Nevermind. If you can mention the name of the example in the figures' captions, it will suffice. As of now, I do not see any changes in the captions of the figures.
