### 2024-08-25 #6916 @prashjha
@ athrasher7, on Joss paper not being updated, the problem is that your updates are in the Docs branch, whereas the branch the paper is being pulled from is the 'main' branch.

There are two ways to fix this: either you update the 'main' branch, pulling Docs into it, or I can replace 'main' with 'Docs' here so that the paper is compiled from the 'Docs' branch. Let me know.
