### 2024-08-26 #6916 @athrasher7
@prashjha I'd prefer that we pull the paper from the Docs branch, as we are trying to get the main branch as really just the base code. In which case, I'd rather not have the paper folder in the main branch at all. Let me know when you've switched it here so it pulls from Docs, and then I will just delete the paper folder on main.
