### 2024-08-28 #6916 @prashjha
Hi @athrasher7, You have been pinging me quite a lot. Please note that this is voluntary work, and if you are not too happy with the progress, feel free to contact the JOSS Editor. Thanks.
