### 2024-08-28 5d06d74
preventing divide by zero in aberration
