### 2024-09-03 #6916 @athrasher7
This is wonderful news! We certainly appreciate all the work you've done to help us get SONIC out into the world. Many thanks from all of us at the Space Exploration Analysis Laboratory.
