### 2024-09-03 #6916 @kyleniemeyer
@athrasher7 we have just a few things to correct before publishing:
- The Zenodo archive is missing some of your coauthors - normally the author list should match the paper's author list.
- Some of the references are not formatting properly - please take a look. Also, when you have a citation to multiple references, please use a single command but separated by a semicolon: `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
