### 2024-09-04 8ac0c85
Update paper.md
Fixing some references.
