### 2024-09-04 b432fc7
Update paper.md
fixing reference
