### 2024-09-04 98a7b1a
Update paper.bib
fixing refs
