+++ paper/paper.bib
@@ -45,7 +45,7 @@
   url = {https://github.com/dstndstn/astrometry.net}
 }
 
-@article{lang2010,
+@article{lang:2010,
   author = {D. Lang and D.W. Hogg and K. Mierle and M. Blanton and S. Roweis},
   title = {Astrometry.net: Blind astrometric calibration of arbitrary astronomical images},
   journal = {The Astronomical Journal},
@@ -55,7 +55,7 @@
   doi = {10.1088/0004-6256/139/5/1782}
 }
 
-@article{sangwine2016, 
+@article{sangwine:2016, 
   author = {Stephen J. Sangwine and Eckhard Hitzer}, 
   title = {Clifford Multivector Toolbox (for {MATLAB})},
   year = 2016, month = {apr}, 
@@ -65,7 +65,7 @@
   doi = {10.1007/s00006-016-0666-x}
 }
 
-@article{henry2023,
+@article{henry:2023,
   author={Henry, Sébastien and Christian, John A.},
   title={Absolute Triangulation Algorithms for Space Exploration},
   volume={46},
@@ -77,7 +77,7 @@
   doi={10.2514/1.g006989}
 }
 
-@article{christian2021,
+@article{christian:2021,
   author={Christian, John A.},
   title={A Tutorial on Horizon-Based Optical Navigation and Attitude Determination With Space Imaging Systems},
   volume={9},
@@ -87,7 +87,7 @@
   doi={10.1109/ACCESS.2021.3051914}
 }
 
-@book{hartley2003,
+@book{hartley:2003,
   author    = {Hartley, R. and Zisserman, A.},
   title     = {Multiple View Geometry in Computer Vision},
   year      = {2003},
@@ -95,7 +95,7 @@
   doi       = {10.1017/CBO9780511811685}
 }
 
-@book{semple1952,
+@book{semple:1952,
   author    = {J. G. Semple and G. T. Kneebone},
   title     = {Algebraic Projective Geometry},
   year      = {1952},
@@ -103,7 +103,7 @@
   doi       = {10.2307/3608975}
 }
 
-@book{richtergebert2011,
+@book{richtergebert:2011,
   author    = {J. Richter-Gebert},
   title     = {Perspective on Projective Geometry: A Guided Tour Through Real and Complex Geometry},
   year      = {2011},
@@ -111,14 +111,14 @@
   doi       = {10.1007/978-3-642-17286-1}
 }
 
-@book{lengyel2024,
+@book{lengyel:2024,
   author    = {E. Lengyel},
   title     = {Projective Geometry Algebra Illuminated},
   year      = {2024},
   publisher = {Terathon Software, LLC},
 }
 
-@article{mortari2004,
+@article{mortari:2004,
   author = {D. Mortari and M. A. Samaan and C. Bruccoleri and J. L. Junkins},
   title = {The Pyramid Star Identification Technique},
   journal = {Navigation: The Journal of the Institute of Navigation},
