### 2024-09-04 #6916 @athrasher7
1. The Zenodo Author list is updated
2. I believe all the references have been updated as you mentioned.

Please let me know if there is anything else I need to do! Much appreciated.
